Move vehicle sounds update into Ui library
