+++ src/openrct2-ui/ride/VehicleSounds.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace OpenRCT2::Audio
+{
+    void UpdateVehicleSounds();
+}
