+++ src/openrct2/windows/Intent.h
@@ -43,6 +43,7 @@ enum IntentAction
     INTENT_ACTION_UPDATE_CASH,
     INTENT_ACTION_UPDATE_BANNER,
     INTENT_ACTION_UPDATE_RESEARCH,
+    INTENT_ACTION_UPDATE_VEHICLE_SOUNDS,
     INTENT_ACTION_TRACK_DESIGN_REMOVE_PROVISIONAL,
     INTENT_ACTION_TRACK_DESIGN_RESTORE_PROVISIONAL,
     INTENT_ACTION_SET_MAP_TOOLTIP,
